requirements.txt
six
python-dateutil>=2.2
python-swiftclient>=2.0.3
ipython[notebook]>=2.0.0
